+++ HISTORY.md
@@ -6,6 +6,9 @@ No advantage over using file_get_contents() and parsing contents.
 
 Instead using PHP gererator function to split processing to discrete chunks so that only the current row is in memory.
 
+# 2019-12-22 Check that EPC and HIP graphics are in correct index
+- Add test for caption is not 'EPC' or 'HIP' for property images
+
 # 2019-12-20 Check media filenames
 - Add test media files are in the format <AGENT_REF>_<MEDIATYPE>_<n>.<file extension>